repair item update

+++ services/item.service.js
@@ -136,6 +136,7 @@ module.exports = {
         .populate({ path: "categoryId", select: "id name" });
 
       if (req.files.length == item.imageId.length) {
+        console.log("masuk 1");
         for (let i = 0; i < item.imageId.length; i++) {
           let imageUpdate = await Image.findOne({ _id: item.imageId[i]._id });
           await fs.unlink(path.join(`public/${imageUpdate.imageUrl}`));
@@ -151,7 +152,11 @@ module.exports = {
 
         req.flash("alertMessage", "success update item");
         req.flash("alertStatus", "success");
-      } else if (req.files.length != item.imageId.length) {
+      } else if (
+        req.files.length != item.imageId.length &&
+        req.files.length > 0
+      ) {
+        console.log("masuk 2");
         for (let i = 0; i < item.imageId.length; i++) {
           let imageUpdate = await Image.findOne({ _id: item.imageId[i]._id });
           await fs.unlink(path.join(`public/${imageUpdate.imageUrl}`));
